Update TODO, quiet_fatal quandary solved.
The LDPO classes now get passed a reference to the parent LDP instance
when they get created and they can use it to look for various
configuration information.  This assumes that they're made to play
together (or are at least familiar with each other), it would be
problematic for an LDPO to check its parent for "has_bark" and then
decide to be a tree instead of a dog.  Caveat emptor...